Move contextprops into src/core
