+++ test/unit/preact/test-slot.js
@@ -15,11 +15,7 @@
  */
 
 import * as Preact from '../../../src/preact/index';
-import {
-  CanPlay,
-  CanRender,
-  LoadingProp,
-} from '../../../src/preact/contextprops';
+import {CanPlay, CanRender, LoadingProp} from '../../../src/core/contextprops';
 import {Slot} from '../../../src/preact/slot';
 import {WithAmpContext} from '../../../src/preact/context';
 import {mount} from 'enzyme';
